+++ pkg/engine/validation.go
@@ -16,9 +16,10 @@ import (
 	"github.com/go-logr/logr"
 	gojmespath "github.com/jmespath/go-jmespath"
 	"github.com/kyverno/kyverno/pkg/engine/response"
-	"github.com/kyverno/kyverno/pkg/engine/utils"
+	engineutils "github.com/kyverno/kyverno/pkg/engine/utils"
 	"github.com/kyverno/kyverno/pkg/engine/validate"
 	"github.com/kyverno/kyverno/pkg/engine/variables"
+	"github.com/kyverno/kyverno/pkg/utils"
 	"k8s.io/apimachinery/pkg/apis/meta/v1/unstructured"
 	"sigs.k8s.io/controller-runtime/pkg/log"
 )
@@ -179,7 +180,7 @@ func newValidator(log logr.Logger, ctx *PolicyContext, rule *kyverno.Rule) *vali
 
 func newForeachValidator(foreach *kyverno.ForEachValidation, rule *kyverno.Rule, ctx *PolicyContext, log logr.Logger) *validator {
 	ruleCopy := rule.DeepCopy()
-	anyAllConditions, err := common.ToMap(foreach.AnyAllConditions)
+	anyAllConditions, err := utils.ToMap(foreach.AnyAllConditions)
 	if err != nil {
 		log.Error(err, "failed to convert ruleCopy.Validation.ForEachValidation.AnyAllConditions")
 	}
@@ -198,16 +199,16 @@ func newForeachValidator(foreach *kyverno.ForEachValidation, rule *kyverno.Rule,
 
 func (v *validator) validate() *response.RuleResponse {
 	if err := v.loadContext(); err != nil {
-		return ruleError(v.rule, utils.Validation, "failed to load context", err)
+		return ruleError(v.rule, engineutils.Validation, "failed to load context", err)
 	}
 
 	preconditionsPassed, err := checkPreconditions(v.log, v.ctx, v.anyAllConditions)
 	if err != nil {
-		return ruleError(v.rule, utils.Validation, "failed to evaluate preconditions", err)
+		return ruleError(v.rule, engineutils.Validation, "failed to evaluate preconditions", err)
 	}
 
 	if !preconditionsPassed && v.ctx.Policy.GetSpec().ValidationFailureAction != kyverno.Audit {
-		return ruleResponse(v.rule, utils.Validation, "preconditions not met", response.RuleStatusSkip)
+		return ruleResponse(v.rule, engineutils.Validation, "preconditions not met", response.RuleStatusSkip)
 	}
 
 	if v.deny != nil {
@@ -216,14 +217,14 @@ func (v *validator) validate() *response.RuleResponse {
 
 	if v.pattern != nil || v.anyPattern != nil {
 		if err = v.substitutePatterns(); err != nil {
-			return ruleError(v.rule, utils.Validation, "variable substitution failed", err)
+			return ruleError(v.rule, engineutils.Validation, "variable substitution failed", err)
 		}
 
 		ruleResponse := v.validateResourceWithRule()
 		if isUpdateRequest(v.ctx) {
 			priorResp, err := validateOldObject(v.log, v.ctx, v.rule)
 			if err != nil {
-				return ruleError(v.rule, utils.Validation, "failed to validate old object", err)
+				return ruleError(v.rule, engineutils.Validation, "failed to validate old object", err)
 			}
 
 			if isSameRuleResponse(ruleResponse, priorResp) {
@@ -241,14 +242,14 @@ func (v *validator) validate() *response.RuleResponse {
 
 func (v *validator) validateForEach() *response.RuleResponse {
 	if err := v.loadContext(); err != nil {
-		return ruleError(v.rule, utils.Validation, "failed to load context", err)
+		return ruleError(v.rule, engineutils.Validation, "failed to load context", err)
 	}
 
 	preconditionsPassed, err := checkPreconditions(v.log, v.ctx, v.anyAllConditions)
 	if err != nil {
-		return ruleError(v.rule, utils.Validation, "failed to evaluate preconditions", err)
+		return ruleError(v.rule, engineutils.Validation, "failed to evaluate preconditions", err)
 	} else if !preconditionsPassed && v.ctx.Policy.GetSpec().ValidationFailureAction != kyverno.Audit {
-		return ruleResponse(v.rule, utils.Validation, "preconditions not met", response.RuleStatusSkip)
+		return ruleResponse(v.rule, engineutils.Validation, "preconditions not met", response.RuleStatusSkip)
 	}
 
 	foreachList := v.rule.Validation.ForEachValidation
@@ -278,10 +279,10 @@ func (v *validator) validateForEach() *response.RuleResponse {
 	}
 
 	if applyCount == 0 {
-		return ruleResponse(v.rule, utils.Validation, "rule skipped", response.RuleStatusSkip)
+		return ruleResponse(v.rule, engineutils.Validation, "rule skipped", response.RuleStatusSkip)
 	}
 
-	return ruleResponse(v.rule, utils.Validation, "rule passed", response.RuleStatusPass)
+	return ruleResponse(v.rule, engineutils.Validation, "rule passed", response.RuleStatusPass)
 }
 
 func (v *validator) validateElements(foreach *kyverno.ForEachValidation, elements []interface{}, elementScope bool) (*response.RuleResponse, int) {
@@ -295,7 +296,7 @@ func (v *validator) validateElements(foreach *kyverno.ForEachValidation, element
 		ctx := v.ctx.Copy()
 		if err := addElementToContext(ctx, e, i, elementScope); err != nil {
 			v.log.Error(err, "failed to add element to context")
-			return ruleError(v.rule, utils.Validation, "failed to process foreach", err), applyCount
+			return ruleError(v.rule, engineutils.Validation, "failed to process foreach", err), applyCount
 		}
 
 		foreachValidator := newForeachValidator(foreach, v.rule, ctx, v.log)
@@ -308,17 +309,17 @@ func (v *validator) validateElements(foreach *kyverno.ForEachValidation, element
 			continue
 		} else if r.Status != response.RuleStatusPass {
 			msg := fmt.Sprintf("validation failure: %v", r.Message)
-			return ruleResponse(v.rule, utils.Validation, msg, r.Status), applyCount
+			return ruleResponse(v.rule, engineutils.Validation, msg, r.Status), applyCount
 		}
 
 		applyCount++
 	}
 
-	return ruleResponse(v.rule, utils.Validation, "", response.RuleStatusPass), applyCount
+	return ruleResponse(v.rule, engineutils.Validation, "", response.RuleStatusPass), applyCount
 }
 
 func addElementToContext(ctx *PolicyContext, e interface{}, elementIndex int, elementScope bool) error {
-	data, err := common.ToMap(e)
+	data, err := utils.ToMap(e)
 	if err != nil {
 		return err
 	}
@@ -359,24 +360,24 @@ func (v *validator) validateDeny() *response.RuleResponse {
 	anyAllCond := v.deny.GetAnyAllConditions()
 	anyAllCond, err := variables.SubstituteAll(v.log, v.ctx.JSONContext, anyAllCond)
 	if err != nil {
-		return ruleError(v.rule, utils.Validation, "failed to substitute variables in deny conditions", err)
+		return ruleError(v.rule, engineutils.Validation, "failed to substitute variables in deny conditions", err)
 	}
 
 	if err = v.substituteDeny(); err != nil {
-		return ruleError(v.rule, utils.Validation, "failed to substitute variables in rule", err)
+		return ruleError(v.rule, engineutils.Validation, "failed to substitute variables in rule", err)
 	}
 
 	denyConditions, err := common.TransformConditions(anyAllCond)
 	if err != nil {
-		return ruleError(v.rule, utils.Validation, "invalid deny conditions", err)
+		return ruleError(v.rule, engineutils.Validation, "invalid deny conditions", err)
 	}
 
 	deny := variables.EvaluateConditions(v.log, v.ctx.JSONContext, denyConditions)
 	if deny {
-		return ruleResponse(v.rule, utils.Validation, v.getDenyMessage(deny), response.RuleStatusFail)
+		return ruleResponse(v.rule, engineutils.Validation, v.getDenyMessage(deny), response.RuleStatusFail)
 	}
 
-	return ruleResponse(v.rule, utils.Validation, v.getDenyMessage(deny), response.RuleStatusPass)
+	return ruleResponse(v.rule, engineutils.Validation, v.getDenyMessage(deny), response.RuleStatusPass)
 }
 
 func (v *validator) getDenyMessage(deny bool) string {
@@ -480,22 +481,22 @@ func (v *validator) validatePatterns(resource unstructured.Unstructured) *respon
 				v.log.V(3).Info("validation error", "path", pe.Path, "error", err.Error())
 
 				if pe.Skip {
-					return ruleResponse(v.rule, utils.Validation, pe.Error(), response.RuleStatusSkip)
+					return ruleResponse(v.rule, engineutils.Validation, pe.Error(), response.RuleStatusSkip)
 				}
 
 				if pe.Path == "" {
-					return ruleResponse(v.rule, utils.Validation, v.buildErrorMessage(err, ""), response.RuleStatusError)
+					return ruleResponse(v.rule, engineutils.Validation, v.buildErrorMessage(err, ""), response.RuleStatusError)
 				}
 
-				return ruleResponse(v.rule, utils.Validation, v.buildErrorMessage(err, pe.Path), response.RuleStatusFail)
+				return ruleResponse(v.rule, engineutils.Validation, v.buildErrorMessage(err, pe.Path), response.RuleStatusFail)
 			}
 
-			return ruleResponse(v.rule, utils.Validation, v.buildErrorMessage(err, pe.Path), response.RuleStatusError)
+			return ruleResponse(v.rule, engineutils.Validation, v.buildErrorMessage(err, pe.Path), response.RuleStatusError)
 		}
 
 		v.log.V(4).Info("successfully processed rule")
 		msg := fmt.Sprintf("validation rule '%s' passed.", v.rule.Name)
-		return ruleResponse(v.rule, utils.Validation, msg, response.RuleStatusPass)
+		return ruleResponse(v.rule, engineutils.Validation, msg, response.RuleStatusPass)
 	}
 
 	if v.anyPattern != nil {
@@ -505,14 +506,14 @@ func (v *validator) validatePatterns(resource unstructured.Unstructured) *respon
 		anyPatterns, err := deserializeAnyPattern(v.anyPattern)
 		if err != nil {
 			msg := fmt.Sprintf("failed to deserialize anyPattern, expected type array: %v", err)
-			return ruleResponse(v.rule, utils.Validation, msg, response.RuleStatusError)
+			return ruleResponse(v.rule, engineutils.Validation, msg, response.RuleStatusError)
 		}
 
 		for idx, pattern := range anyPatterns {
 			err := validate.MatchPattern(v.log, resource.Object, pattern)
 			if err == nil {
 				msg := fmt.Sprintf("validation rule '%s' anyPattern[%d] passed.", v.rule.Name, idx)
-				return ruleResponse(v.rule, utils.Validation, msg, response.RuleStatusPass)
+				return ruleResponse(v.rule, engineutils.Validation, msg, response.RuleStatusPass)
 			}
 
 			if pe, ok := err.(*validate.PatternError); ok {
@@ -536,11 +537,11 @@ func (v *validator) validatePatterns(resource unstructured.Unstructured) *respon
 
 			v.log.V(4).Info(fmt.Sprintf("Validation rule '%s' failed. %s", v.rule.Name, errorStr))
 			msg := buildAnyPatternErrorMessage(v.rule, errorStr)
-			return ruleResponse(v.rule, utils.Validation, msg, response.RuleStatusFail)
+			return ruleResponse(v.rule, engineutils.Validation, msg, response.RuleStatusFail)
 		}
 	}
 
-	return ruleResponse(v.rule, utils.Validation, v.rule.Validation.Message, response.RuleStatusPass)
+	return ruleResponse(v.rule, engineutils.Validation, v.rule.Validation.Message, response.RuleStatusPass)
 }
 
 func deserializeAnyPattern(anyPattern apiextensions.JSON) ([]interface{}, error) {
